requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
chromedriver-binary-auto==0.3.1
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
outcome==1.2.0
praw==7.7.1
prawcore==2.3.0
PySocks==1.7.1
requests==2.31.0
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.4
update-checker==0.18.0
urllib3==2.0.4
websocket-client==1.6.2
wsproto==1.2.0
